4.21

+++ controllers/bloglist.js
@@ -37,6 +37,17 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+       return response.status(401).json({ error: 'token missing  or invalid' })
+    } 
+    const blog = await Blog.findById(request.params.id)
+    const idToCompare = blog.user
+    if (decodedToken.id.toString() !== idToCompare.toString()) {
+        return response.status(401).json({ error: 'only the user who added the blog can delete it' })
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
